We no longer bind resources to specific stages
It's more complicated, it causes crashes, it takes more memory and ultimately it segregates the resources for no particularly useful reason
It's not like I would like to bind two different textures to the same enum (that would be really confusing), if that were the case then we can create two separate ones
Even in a DX11 model where we'd bind things to different stages the shader would tell us what it needs
This change will also allow us to create tables in DX12 where resources are shared, but that requires more work as we currently separate them

+++ Source/Rendering/Vulkan/CrRenderDevice_vk.cpp
@@ -360,7 +360,7 @@ CrHardwareGPUBufferHandle CrRenderDeviceVulkan::DownloadBufferPS(const ICrHardwa
 {
 	uint32_t stagingBufferSizeBytes = sourceBuffer->GetSizeBytes();
 
-	CrHardwareGPUBufferDescriptor stagingBufferDescriptor(cr3d::BufferUsage::TransferDst, cr3d::MemoryAccess::StagingDownload, (uint32_t)stagingBufferSizeBytes);
+	CrHardwareGPUBufferDescriptor stagingBufferDescriptor(cr3d::BufferUsage::TransferDst, cr3d::MemoryAccess::StagingDownload, stagingBufferSizeBytes);
 	CrHardwareGPUBufferHandle stagingBuffer = CreateHardwareGPUBuffer(stagingBufferDescriptor);
 	CrHardwareGPUBufferVulkan* vulkanStagingBuffer = static_cast<CrHardwareGPUBufferVulkan*>(stagingBuffer.get());
